pixelboost/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
bcrypt==4.3.0
    # via passlib
beanie==1.29.0
    # via -r requirements.in
click==8.1.8
    # via
    #   beanie
    #   uvicorn
colorama==0.4.6
    # via click
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.2.0
    # via pydantic
fastapi==0.115.12
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
lazy-model==0.2.0
    # via beanie
motor==3.7.0
    # via
    #   -r requirements.in
    #   beanie
passlib==1.7.4
    # via -r requirements.in
pydantic==2.11.1
    # via
    #   -r requirements.in
    #   beanie
    #   fastapi
    #   lazy-model
pydantic-core==2.33.0
    # via pydantic
pyjwt==2.10.1
    # via -r requirements.in
pymongo==4.11.3
    # via
    #   -r requirements.in
    #   motor
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via fastapi
toml==0.10.2
    # via beanie
typing-extensions==4.13.0
    # via
    #   anyio
    #   beanie
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
uvicorn==0.34.0
    # via -r requirements.in
